ai_tribal_marketplace/backend/requirements.txt
fastapi
uvicorn
transformers
torch
pillow
sqlalchemy
python-multipart
